set executable for subprocess.run (#71)
The shell execution code changed in #68, `snakemake.shell` was replaced
with `subprocess.run`. This broke the NCBI assemblies processing, and
investigation led to the conclusion it was because `snakemake.shell`
defaults to `/bin/bash` for the shell, while `subprocess.run` uses
`/bin/sh`, and the assembly task uses a `bash` redirect that is a syntax
error for `sh`.

Also fix the shell quoting, since `{:q}` is `snakemake`-specific.