fix: don't upload empty sigs, raise exception with more details for debug later (#77)
On worker: if the signature is empty, don't upload it to S3, and raise
an error for debugging why it failed.

On frontend: Add a `recompute` flag to clean cache and `computed` field
from DB. This doesn't delete any existing objects from S3, still need to
do this outside `wort`.

Clean up logic for verifying if a dataset is already computed, and only
set cache if it is actually computed.

Part of addressing #76